+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @felte/core
 
+## 1.2.0
+
+### Minor Changes
+
+- fd58a47: Add `onReset` to `createField`
+
 ## 1.1.1
 
 ### Patch Changes