+++ packages/solid/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @felte/solid
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [fd58a47]
+  - @felte/core@1.2.0
+
 ## 1.1.1
 
 ### Patch Changes