+++ packages/vanilla/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @felte/vanilla
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [fd58a47]
+  - @felte/core@1.2.0
+
 ## 0.2.1
 
 ### Patch Changes